fix(omnigraph): make the cluster-catalog logic importable without AWS
The new test could not even be collected in CI:

    botocore.exceptions.NoRegionError: You must specify a region.
    ERROR tests/…/test_cluster_policies.py
    Interrupted: 1 error during collection

Importing `data_tier` pulls in pulumi_aws and the EKS/S3 components,
which want credentials and a region at import time. It passed locally
only because a region happened to be configured, which is exactly the
kind of difference that should not decide whether a test runs.

Move the pure cluster-catalog logic — the graph ids, the schema and
policy filenames, `code_graph_id`, `build_cluster_graphs` and
`build_cluster_policies` — into `cluster_config.py`, which imports
nothing beyond `hashlib` and `re`. `storage.py` was split out for the
same reason and is tested the same way.

Worth having independently of the test: `code_graph_id` is a
byte-for-byte shared contract with agent-kit's client-side
`witan_code.config.graph_id`, and until now nothing could exercise it
without standing up AWS.

Behaviour-neutral, and checked rather than assumed: `pulumi preview`
against CI is identical to the previous commit (~1 update, +-2 replace,
3 changes, 42 unchanged), the full omnigraph suite passes (61), and the
new test now collects and passes with every AWS variable unset and an
empty HOME — the CI condition that caught this.
